requirements.txt
Pillow~=7.2.0
torch~=1.6
torchvision~=0.7
matplotlib
flask
sparklines
pandas~=1.1.1
numpy~=1.19.1
altair
cython~=0.29
tqdm~=4.50.0
attrs~=20.1.0
setuptools~=49.6.0
jinja2~=2.11.2
pytorch-ignite
pyocclient
joblib
pycocotools
opencv-python
tensorboard~=2.3.0
tensorboardX~=2.1